requirements.txt
ansi2html==1.8.0
asttokens==2.0.5
Babel==2.10.3
backcall==0.2.0
blinker==1.5
Brotli==1.0.9
certifi==2022.6.15
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
dash==2.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
Flask==2.1.3
Flask-Babel==2.0.0
Flask-Compress==1.12
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
Flask-WhooshAlchemy==0.56
Flask-WTF==1.0.1
flup==1.0.3
greenlet==1.1.2
idna==3.3
ipykernel==6.15.1
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyter-dash==0.4.2
lamson==1.3.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.23.1
packaging==21.3
pandas==1.4.3
parso==0.8.3
pbr==5.9.0
pickleshare==0.7.5
plotly==5.9.0
prompt-toolkit==3.0.30
psutil==5.9.1
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.9
pyTelegramBotAPI==4.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
pywin32==304
pyzmq==23.2.0
requests==2.28.1
retrying==1.3.3
six==1.16.0
SQLAlchemy==1.4.39
sqlalchemy-migrate==0.13.0
sqlparse==0.4.2
stack-data==0.3.0
Tempita==0.5.2
tenacity==8.0.1
tornado==6.2
traitlets==5.3.0
urllib3==1.26.10
wcwidth==0.2.5
Werkzeug==2.1.2
Whoosh==2.7.4
WTForms==3.0.1
